+++ ecommerce/src/main/java/com/ultra/ecommerce/repository/ProductRepository.java
@@ -4,9 +4,10 @@ import com.ultra.ecommerce.entity.Product;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
     Optional<Product> findByName(String name);
 }
